pyproject.toml
[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88
src_paths = "manser,test"

[tool.poetry]
name = "manser"
version = "0.1.0"
description = "Manser - Rest API to find update of mangas chapters"
authors = ["pavkazzz <me@pavkazzz.ru>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.55.1"
aiohttp = "^3.6.2"
cssselect = "^1.1.0"
lxml = "^4.5.1"
chardet = "^3.0.4"
uvicorn = "^0.11.5"
orjson = "^3.0.2"
typer = "^0.2.1"
lsm-db = "^0.6.4"
starlette_exporter = "^0.4.0"
Telethon = "^1.14.0"
aiocache = "^0.11.1"
aresponses = "^2.0.0"
aioitertools = "^0.7.0"
aiohttp_socks = "^0.4.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
requests = "^2.23.0"
black = "^19.10b0"
maturin = "^0.8.1"
setuptools = "^47.1.1"
isort = "^5.0.7"
pytest-asyncio = "^0.14.0"
asyncio-socks5 = "^0.2.11"
httpx = "^0.13.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
